# api/docs/lefax.v0.0.1-openapi.yml
openapi: 3.0.2
info:
  title: LeFax API
  version: 0.0.1
  description: ""
  contact:
    name: Ekoulé Maneng
    email: ekoulemaneng@gmail.com
    url: www.lefax.cm
  license:
      name: OAPI 2015 Licence
...
servers:
  - url: 'http://localhost:8080/api/v0.0.1'
    description: development server 
  - url: 'https://letax-api.ekoulemaneng.repl.co/api/v0.0.1/'
    description: production server
paths:
...
  /admins:
    post:
      tags:
        - Admins
      summary: Create an admin account
      operationId: CreateAdmin
      requestBody:
        description: Created admin object
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
        required: true
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdResponse'
    get:
      tags:
        - Admins
      summary: Get admin account details
      operationId: GetAdminDetails
      security:
        - api_key: []
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDetailsResponse'
    delete:
      tags:
        - Admins
      summary: Delete admin account
      operationId: DeleteAdmin
      security:
        - api_key: []
      requestBody:
        description: Create an object
        content:
          application/x-www-form-urlencoded:
            schema:
             $ref: '#/components/schemas/DeleteUserRequest'
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  details:
                    type: string
                    default: AdminSuccessfullyDeleted
  /admins/login:
    post:
      tags:
        - Admins
      summary: Log in admin account to get api token
      operationId: LogInAdmin
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CredentialsUserRequest'
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogInUserResponse'
  /admins/name:
    put:
      tags:
        - Admins
      summary: Update admin name
      operationId: UpdateAdminName
      security:
        - api_key: []
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/UpdateUserNameRequest'
        required: true
      responses:
        200:
          description: Admin name updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    description: Admin Id
                  name:
                    type: string
                    description: Admin name
  /admins/password:
    put:
      tags:
        - Admins
      summary: Modify admin password
      operationId: ModifyAdminPassword
      security:
        - api_key: []    
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/UpdateUserPasswordRequest'
        required: true
      responses:
        200:
          description: Admin password updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  details:
                    type: string
                    default: PasswordSuccessfullyModified 
...
  /students/password:
    put:
      tags:
        - Students
      summary: Update admin password
      operationId: UpdateAdminPassword
      security:
        - api_key: []    
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/UpdateUserPasswordRequest'
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  details:
                    type: string
                    default: PasswordSuccessfullyModified 
...
  /countries:
    post:
      tags:
        - Countries
      summary: Add a country
      operationId: AddCountry
      security:
        - api_key: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Country name
                code:
                  type: string
                  description: Country code
                image:
                  type: string
                  format: binary
                  description: Country image
        required: true
      responses:
        200:
          description: A new country added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdResponse'
    get:
      tags:
        - Countries
      summary: Get all countries
      operationId: GetAllCountries
      responses:
        200: 
          description: List of countries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CountryDetailsResponse'
  /countries/search:
    get:
      tags:
        - Countries
      summary: Get a country by code
      operationId: GetCountryByCode
      parameters:
        - in: query
          name: code
          schema:
            type: string
          description: Country code
      responses:
        200:
          description: Country found
          content: 
            application/json:
              schema: 
                $ref: '#/components/schemas/CountryDetailsResponse'
  /countries/{countryId}:
    get:
      tags:
        - Countries
      summary: Get a country by id
      operationId: GetCountryById
      parameters:
        - in: path
          name: countryId
          schema:
            type: string
          description: Country Id
          required: true
      responses:
        200: 
          description: Country found
          content: 
            application/json:
              schema: 
                $ref: '#/components/schemas/CountryDetailsResponse'
    delete:
      tags:
        - Countries
      summary: Delete a country
      operationId: DeleteCountry
      security:
        - api_key: []
      parameters:
        - in: path
          name: countryId
          schema:
            type: string
          description: Country Id
          required: true
      responses:
        200:
          description: Country deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  details:
                    type: string
                    description: Error message
                    default: CountrySuccessfullyDeleted
  /countries/{countryId}/infos:  
    put:
      tags:
        - Countries
      summary: Update country infos (name and code)
      operationId: UpdateCountryInfos
      security:
        - api_key: []  
      parameters:
        - in: path
          name: countryId
          schema:
            type: string
          description: Country id
          required: true
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Country name
                code:
                  type: string
                  description: Country code 
      responses:
        200:
          description: Country updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountryDetailsResponse'
  /countries/{countryId}/image:  
    put:
      tags:
        - Countries
      summary: Upload a new country image 
      operationId: UpdateCountryImage
      security:
        - api_key: []  
      parameters:
        - in: path
          name: countryId
          schema:
            type: string
          description: Country id
          required: true
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                  description: Country image
      responses:
        200:
          description: Country updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountryImageResponse'
...
components:
    schemas:
...
      CreateUserRequest:
        type: object
        properties:
          name:
            type: string
            description: User name
          email:
            type: string
            description: User email
          password:
            type: string
            description: User password
          repeat_password:
            type: string
            description: Rewrite the same password 
      CredentialsUserRequest:
        type: object
        properties:
          email:
            type: string
            description: User email
          password:
            type: string
            description: User password
      UpdateUserNameRequest:
        type: object
        properties:
          name:
            type: string
            description: User name
      UpdateUserPasswordRequest:
        type: object
        properties:
          oldPassword:
            type: string
            description: Current user password
          newPassword:
            type: string
            description: New user password
      DeleteUserRequest:
        type: object
        properties:
          password:
            type: string
            description: User password
...
      IdResponse:
        type: object
        properties:
          id:
            type: string
            description: User/thing id
      LogInUserResponse:
        type: object
        properties:
          id:
            type: string
            description: User id
          token:
            type: string
            description: Authentication token
      UserDetailsResponse:
        type: object
        properties:
          id:
            type: string
            description: User id
          name:
            type: string
            description: User name
          email:
            type: string
            description: User email
      AdminDetailsResponse:
        allOf: 
          - $ref: '#/components/schemas/UserDetailsResponse'
          - type: object
            properties:
              role:
                type: string
                default: admin
...
      CountryDetailsResponse: 
        type: object
        properties:
          id:
            type: string
            description: Country id
          name:
            type: string
            description: Country name
          code:
            type: string
            description: Country code
          image:
            type: string
            description: Image url
      CountryImageResponse: 
        type: object
        properties:
          id:
            type: string
            description: Country id
          image:
            type: string
            description: Image url
...
    securitySchemes:
      api_key:
        type: apiKey
        name: authorization
        in: header